Extracted interference strategy
