+++ ModelTelecomNetConsole/AccurateMathematic.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ModelTelecomNetConsole
+{
+    public abstract class AccurateMathematic
+    {
+        // Accuracy of built-in PI constant is not enough.
+        protected const double PI = 3.14159265358979323846;
+    }
+}
